+++ lib/slack_message.rb
@@ -11,19 +11,15 @@ module SlackMessage
     configuration.configure(&block)
   end
 
-  def self.user_id_for(email) # spooky undocumented public method 👻
-    Api::user_id_for(email)
-  end
-
   def self.post_to(target, as: :default, &block)
     payload = Dsl.new(block).tap do |instance|
       instance.instance_eval(&block)
     end
 
-    profile = Configuration.profile(as, custom_name: payload.custom_bot_name)
-    target  = user_id_for(target) if target =~ /^\S{1,}@\S{2,}\.\S{2,}$/
+    profile = Configuration.profile(as)
+    target  = Api::user_id_for(target, profile) if target =~ /^\S{1,}@\S{2,}\.\S{2,}$/
 
-    Api.post(payload.render, target, profile)
+    Api.post(payload, target, profile)
   end
 
   def self.post_as(profile_name, &block)
@@ -31,15 +27,15 @@ module SlackMessage
       instance.instance_eval(&block)
     end
 
-    profile = Configuration.profile(profile_name, custom_name: payload.custom_bot_name)
+    profile = Configuration.profile(profile_name)
     if profile[:default_channel].nil?
       raise ArgumentError, "Sorry, you need to specify a default_channel for profile #{profile_name} to use post_as"
     end
 
     target  = profile[:default_channel]
-    target  = user_id_for(target) if target =~ /^\S{1,}@\S{2,}\.\S{2,}$/
+    target  = Api::user_id_for(target, profile) if target =~ /^\S{1,}@\S{2,}\.\S{2,}$/
 
-    Api.post(payload.render, target, profile)
+    Api.post(payload, target, profile)
   end
 
   def self.build(&block)