+++ lib/slack_message/api.rb
@@ -3,13 +3,11 @@ require 'net/https'
 require 'json'
 
 class SlackMessage::Api
-  def self.user_id_for(email)
-    token = SlackMessage.configuration.api_token
-
+  def self.user_id_for(email, profile)
     uri = URI("https://slack.com/api/users.lookupByEmail?email=#{email}")
     request = Net::HTTP::Get.new(uri).tap do |req|
-      req['Authorization']  = "Bearer #{token}"
-      req['Content-type']   = "application/json"
+      req['Authorization']  = "Bearer #{profile[:api_token]}"
+      req['Content-type']   = "application/json; charset=utf-8"
     end
 
     response = Net::HTTP.start(uri.hostname, uri.port, use_ssl: true) do |http|
@@ -38,26 +36,37 @@ class SlackMessage::Api
   end
 
   def self.post(payload, target, profile)
-    profile[:url] = profile[:url]
-
-    uri     = URI.parse(profile[:url])
     params  = {
       channel: target,
-      username: profile[:name],
-      blocks: payload
+      username: payload.custom_bot_name || profile[:name],
+      blocks: payload.render,
+      text: payload.notification_text,
     }
 
-    response = execute_post_form(uri, params, profile[:handle])
+    icon = payload.custom_bot_icon || profile[:icon]
+    if icon =~ /^:\w+:$/
+      params[:icon_emoji] = icon
+    elsif icon =~ /^(https?:\/\/)?[0-9a-z]+\.[-_0-9a-z]+/ # very naive regex, I know. it'll be fine.
+      params[:icon_url] = icon
+    elsif !(icon.nil? || icon == '')
+      raise ArgumentError, "Couldn't figure out icon '#{icon}'. Try :emoji: or a URL."
+    end
+
+    response = post_message(profile, params)
+    body  = JSON.parse(response.body)
+    error = body.fetch("error", "")
 
     # let's try to be helpful about error messages
-    if response.body == "invalid_token"
-      raise "Couldn't send slack message because the URL for profile '#{profile[:handle]}' is wrong."
-    elsif response.body == "channel_not_found"
+    if ["token_revoked", "token_expired", "invalid_auth", "not_authed"].include?(error)
+      raise "Couldn't send slack message because the API key for profile '#{profile[:handle]}' is wrong."
+    elsif ["no_permission", "ekm_access_denied"].include?(error)
+      raise "Couldn't send slack message because the API key for profile '#{profile[:handle]}' isn't allowed to post messages."
+    elsif error == "channel_not_found"
       raise "Tried to send Slack message to non-existent channel or user '#{target}'"
-    elsif response.body == "missing_text_or_fallback_or_attachments"
+    elsif error == "invalid_arguments"
       raise "Tried to send Slack message with invalid payload."
     elsif response.code == "302"
-      raise "Got 302 response while posting to Slack. Check your webhook URL for '#{profile[:handle]}'."
+      raise "Got 302 response while posting to Slack. Check your API key for profile '#{profile[:handle]}'."
     elsif response.code != "200"
       raise "Got an error back from the Slack API (HTTP #{response.code}):\n#{response.body}"
     end
@@ -65,7 +74,19 @@ class SlackMessage::Api
     response
   end
 
-  def self.execute_post_form(uri, params, _profile)
-    Net::HTTP.post_form uri, { payload: params.to_json }
+  # mostly test harness
+  def self.post_message(profile, params)
+    uri = URI("https://slack.com/api/chat.postMessage")
+    request = Net::HTTP::Post.new(uri).tap do |req|
+      req['Authorization']  = "Bearer #{profile[:api_token]}"
+      req['Content-type']   = "application/json; charset=utf-8"
+      req.body = params.to_json
+    end
+
+    Net::HTTP.start(uri.hostname, uri.port, use_ssl: true) do |http|
+      http.request(request)
+    end
   end
+
+  private_class_method :post_message
 end